# revert: perf: do not detect cycles in dependencies